requirements.txt
torch
torchvision
tensorboard
scipy
numpy
matplotlib
opencv_python
tqdm
Pillow
h5py
labelme==3.16.7